requirements.txt
nats-py
nats-py[nkeys]
asyncio